refactor: Move router to own folder and break down into componentes
each component will handle the type of route
