+++ src/routes/index.tsx
@@ -1,0 +1,13 @@
+import { NavigationContainer } from "@react-navigation/native";
+import { createStackNavigator } from "@react-navigation/stack";
+import StackRoutes from "./stack.routes";
+
+const Router = () => {
+  return (
+    <NavigationContainer>
+      <StackRoutes />
+    </NavigationContainer>
+  );
+};
+
+export default Router;
