+++ src/routes/stack.routes.tsx
@@ -1,0 +1,54 @@
+import { createStackNavigator } from "@react-navigation/stack";
+import { RootStackParamList } from "@type/routeType";
+import Login from "@pages/Login";
+import SignUp from "@pages/SignUp";
+import EmailRequest from "@pages/ResetPassword/1-EmailRequest";
+import CodeRequest from "@pages/ResetPassword/2-CodeRequest";
+import NewPassword from "@pages/ResetPassword/3-NewPassword";
+import Policy from "@pages/Policy";
+import Team from "@pages/Team";
+import LastPeriod from "@pages/MenstrualInitialSettings/2-LastPeriod";
+import CycleDuration from "@pages/MenstrualInitialSettings/3-CycleDuration";
+
+const Stack = createStackNavigator<RootStackParamList>();
+
+const StackRoutes = () => {
+  return (
+    <Stack.Navigator
+      screenOptions={{
+        headerShown: false
+      }}
+    >
+      {/* If user has no account or has logged out */}
+      <Stack.Group>
+        <Stack.Screen name="Login" component={Login} />
+
+        {/* Account creation plus its initial settings */}
+        <Stack.Group>
+          <Stack.Screen name="SignUp" component={SignUp} />
+          <Stack.Screen name="Policy" component={Policy} />
+          <Stack.Screen name="Team" component={Team} />
+          <Stack.Screen name="LastPeriod" component={LastPeriod} />
+          <Stack.Screen name="CycleDuration" component={CycleDuration} />
+        </Stack.Group>
+
+        {/* Forgotten password */}
+        <Stack.Group>
+          <Stack.Screen name="EmailRequest" component={EmailRequest} />
+          <Stack.Screen
+            options={{ headerShown: true, title: "Redefinir Senha" }}
+            name="CodeRequest"
+            component={CodeRequest}
+          />
+          <Stack.Screen
+            options={{ headerShown: true, title: "Redefinir Senha" }}
+            name="NewPassword"
+            component={NewPassword}
+          />
+        </Stack.Group>
+      </Stack.Group>
+    </Stack.Navigator>
+  );
+};
+
+export default StackRoutes;
